requirements.txt
annotated-types==0.5.0
boto3==1.28.22
botocore==1.31.22
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
jmespath==1.0.1
numpy==1.25.1
opencv-python==4.8.0.74
pydantic==2.1.1
pydantic_core==2.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
typing_extensions==4.7.1
urllib3==1.26.16
